+++ ccm/src/main/java/calendar/service/CalculatorService.java
@@ -3,6 +3,7 @@ package calendar.service;
 import java.sql.Connection;
 import java.sql.Date;
 import java.sql.SQLException;
+import java.util.HashMap;
 
 import calendar.dao.CalendarDao;
 import calendar.dto.Calendar;
@@ -61,5 +62,135 @@ public class CalculatorService {
 		} finally {
 			JdbcUtil.close(conn);
 		}
+	}
+	
+	public void Reset(String memberId) {
+		Connection conn = null;
+		java.sql.Date date = new java.sql.Date(System.currentTimeMillis());
+		System.out.println("sql date : "+date);
+		try {
+			conn = ConnectionProvider.getConnection();
+			conn.setAutoCommit(false);
+			calendarDao.resetCaffeine(date, memberId, conn);
+			
+			conn.commit();
+		}catch (SQLException e) {
+			e.printStackTrace();
+			JdbcUtil.rollback(conn);
+		} finally {
+			JdbcUtil.close(conn);
+		}
+	}
+	
+	
+	/*
+	 * 비동기 메서드
+	 */
+	public HashMap calculateAsync(String memberId, int coffeeNo) {
+		Connection conn = null;
+		Coffee coffee = null;
+		Calendar todaysCaffeine = null;
+		Date date = new Date(System.currentTimeMillis());
+		int caffeine = 0;
+		int calculatedResult = 0;
+		String color;
+		HashMap<String, Object> calcMap = new HashMap();
+		try {
+			conn = ConnectionProvider.getConnection();
+			conn.setAutoCommit(false);
+			
+			double weight = profileDao.getWeight(memberId, conn);
+			
+			coffee = coffeeListDao.selectByCoffeeNo(coffeeNo, conn);
+			caffeine = coffee.getC_CAFFEINE();
+			todaysCaffeine = calendarDao.getTodaysRecord(memberId, date, conn);
+			
+			if (todaysCaffeine != null) {
+				caffeine +=  todaysCaffeine.getCAL_DAILYCF();
+				color = CCMFunctions.ColorFn(caffeine, weight);
+				calendarDao.recordCaffeine(todaysCaffeine.getCAL_DATE(), memberId, caffeine,color, conn);
+
+			} else {
+				color = CCMFunctions.ColorFn(caffeine, weight);
+				calendarDao.createTodaysRecord(date, memberId, caffeine, color, conn);
+			}
+			//========================================
+			Calendar todaysRecord = calendarDao.getTodaysRecord(memberId, date, conn);
+			if(todaysRecord == null) {
+				calculatedResult = 0;
+			} else {
+				calculatedResult = todaysRecord.getCAL_DAILYCF();
+			}
+			calcMap.put("caffeine", calculatedResult);
+			calcMap.put("rda", weight*6.4);
+			calcMap.put("color", color);
+			conn.commit();
+			
+		}catch (Exception e) {
+			e.printStackTrace();
+		}finally {
+			JdbcUtil.close(conn);
+		}
+		return calcMap;
+	}
+	
+	public HashMap<String, Object> resetAsync(String memberId)  {
+		Connection conn = null;
+		Date date = new Date(System.currentTimeMillis());
+		String color;
+		HashMap<String, Object> calcMap = new HashMap();
+		
+		try {
+			conn = ConnectionProvider.getConnection();
+			conn.setAutoCommit(false);
+			
+			double weight = profileDao.getWeight(memberId, conn);
+			calendarDao.resetCaffeine(date, memberId, conn);
+			
+			color = CCMFunctions.ColorFn(0, 1);
+			calcMap.put("caffeine", 0);
+			calcMap.put("rda", weight*6.4);
+			calcMap.put("color", color);
+			conn.commit();
+			
+		}catch (Exception e) {
+			e.printStackTrace();
+		}finally {
+			JdbcUtil.close(conn);
+		}
+		return calcMap;
+	}
+	
+	public HashMap<String, Object> getCalcAsync(String memberId)  {
+		Connection conn = null;
+		Date date = new Date(System.currentTimeMillis());
+		String color;
+		HashMap<String, Object> calcMap = new HashMap();
+		int calculatedResult = 0;
+		try {
+			conn = ConnectionProvider.getConnection();
+			conn.setAutoCommit(false);
+			
+			double weight = profileDao.getWeight(memberId, conn);
+			
+			Calendar todaysRecord = calendarDao.getTodaysRecord(memberId, date, conn);
+			if(todaysRecord == null) {
+				calculatedResult = 0;
+			} else {
+				calculatedResult = todaysRecord.getCAL_DAILYCF();
+			}
+			
+			color = CCMFunctions.ColorFn(calculatedResult, weight);
+			calcMap.put("caffeine", calculatedResult);
+			calcMap.put("rda", weight*6.4);
+			calcMap.put("color", color);
+			conn.commit();
+			
+		}catch (Exception e) {
+			e.printStackTrace();
+		}finally {
+			JdbcUtil.close(conn);
+		}
+		return calcMap;
 	}
 }
