+++ ccm/src/main/webapp/WEB-INF/config/controller.properties
@@ -1,3 +1,6 @@
+#test
+/test.do = global.handler.TestHandler
+
 /domain/command.do = global.handler.NullHandler
 /LoginHandler.do = member.handler.LoginHandler
 /JoinMemberHandler.do = member.handler.JoinMemberHandler
@@ -6,18 +9,21 @@
 /CheckHandler.do = member.handler.CheckHandler
 /views/screens/CheckHandler.do = member.handler.CheckHandler
 
+#Calculator
 /calculate.do = calendar.handler.CalculatorHandler
 /reset.do = calendar.handler.ResetCalculationHandler
+
+#Favorite list
 /addfav.do = favorite.handler.AddFavoriteHandler
 /deletefav.do = favorite.handler.DeleteFavHandler
-/main.do = global.handler.MainPageHandler
 
-/test.do = global.handler.TestHandler
-/mypage.do = mypage.handler.MypageHandler
 
 /CustomBoardAddHandler.do=customRecipe.handler.CustomBoardAddHandler
-/CustomBoardViewHandler.do=customRecipe.handler.CustomBoardViewHandler
 /CustomBoardListHandler.do=customRecipe.handler.CustomBoardListHandler
-/main.do = global.handler.MainPageHandler
 
+# Page viewer
+/CustomBoardViewHandler.do=customRecipe.handler.CustomBoardViewHandler
+/main.do = global.handler.MainPageHandler
+/coffeelist.do = coffeeList.handler.CoffeeListPageHandler
+/mypage.do = mypage.handler.MypageHandler
